Separate PlayerData from PlayerController
This will allow the engine to own the player data and actual values,
while allowing player controllers (implementations) to have read-only
access to those values.

This is mostly a proof of concept of how to use traits (aka interfaces).
The next step will be to try and implement a controller in a different
library and get it compiled into and used in our game. That will likely
require some system for implementations to "register" themselves under a
specific name, then our command line will accept an array of controller
implementation names and instantiate them dynamically for the game to
operate with.