Add support for dynamic player controllers
This change updates the engine library to take in an array of strings
corresponding to player controllers. We have a player_factory that is
responsible for converting these strings into actual PlayerController
implementations. I was able to add a sample PlayerController to
engine.rs, and update the create_player method to instantiate this type
defined externally. This should indicate we can support types defined in
other libraries.

With that said, it would be great to be able to support types just by
the existence of the library. Right now, we require code changes in the
engine to specifically support player implementations, which feels a bit
clunky.