Begin implementing actions
This change starts work on implementing functional actions. We map the
action enum into an action implementation, which is an engine-internal
class that executes the logic of that action, while we only expose the
action enums for players to use.

TODO: Figure out how to resolve the build error caused by lifetime
issues.